+++ CHANGELOG.md
@@ -9,11 +9,13 @@ omnibus-gitlab repository.
 - Extends rspec to `gitlab-ctl consul`'s helper class, and refactors to avoid
   namespace conflicts.
 - Geo: Use a background WAL receiver and replication slot to improve initial sync
+- Support Redis as an LRU cache
 - Remove usage of gitlab_shell['git_data_directories'] configuration 2003bc5d
 - Set PostgreSQL archive_timeout to 0 by default
 - Added LDAP configuration option for lowercase_usernames
 - Support authorized_keys database lookups with SELinux on CentOS 7.4
 - Add support for generating Let's Encrypt certificates as part of reconfigure
+- Support configuration of Redis Sentinels by persistence class
 - Add support for setting environment variables for Registry (Anthony Dong)
 - Don't attempt to backup non-existent PostgreSQL data file in gitlab-ctl replicate-geo-database
 - Add object storage support for uploads